// src/components/ui/features/button-animation/types.ts
export type EntryDirection = 'right' | 'left' | 'up' | 'down' | 'none'
export type EntryOrder = 'sequential' | 'reverse' | 'center-out'
export type StaggerDirection = 'forward' | 'reverse'